started organizing the pub setting page
