+++ app/lish/[did]/[publication]/dashboard/PublicationDashboard.tsx
@@ -14,7 +14,8 @@ import { useDebouncedEffect } from "src/hooks/useDebouncedEffect";
 import { type NormalizedPublication } from "src/utils/normalizeRecords";
 import { PublicationAnalytics } from "./PublicationAnalytics";
 import { useCanSeePro } from "src/hooks/useEntitlement";
-import { SettingsTab } from "./settings/SettingsTab";
+import { SettingsContent } from "./settings/SettingsContent";
+import { SettingsTiny } from "components/Icons/SettingsTiny";
 
 export default function PublicationDashboard({
   publication,
@@ -68,7 +69,7 @@ export default function PublicationDashboard({
           ),
           controls: null,
         },
-        Subscribers: {
+        Subs: {
           content: (
             <PublicationSubscribers
               showPageBackground={!!record.theme?.showPageBackground}
@@ -89,8 +90,9 @@ export default function PublicationDashboard({
             }
           : {}),
         Settings: {
+          icon: <SettingsTiny />,
           content: (
-            <SettingsTab
+            <SettingsContent
               showPageBackground={!!record.theme?.showPageBackground}
             />
           ),
