+++ app/lish/createPub/UpdatePubForm.tsx
@@ -165,13 +165,13 @@ export const EditPubForm = (props: {
           onToggle={() => setShowInDiscover(!showInDiscover)}
         >
           <div className=" pt-0.5 flex flex-col  text-sm text-tertiary ">
-            <p className="font-bold italic">Show In Discover</p>
-            <p className="text-xs text-tertiary font-normal">
+            <p className="font-bold italic">Make Public</p>
+            <p className="text-sm text-tertiary font-normal">
               Your posts will appear in{" "}
               <a href="/reader" target="_blank">
                 Leaflet Reader
-              </a>
-              . You can change this at any time!
+              </a>{" "}
+              and show up in search and tags. You can change this at any time!
             </p>
           </div>
         </Toggle>
@@ -179,4 +179,3 @@ export const EditPubForm = (props: {
     </form>
   );
 };
-
