+++ components/Icons/SettingsTiny.tsx
@@ -1,0 +1,21 @@
+import { Props } from "./Props";
+
+export const SettingsTiny = (props: Props) => {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M9.29339 0.710621C8.52319 0.574818 7.85522 1.07422 7.63679 1.73718C7.59815 1.85444 7.46828 1.97645 7.25941 2.00195C7.22787 2.0058 7.19635 2.0099 7.16486 2.01424C6.95459 2.04326 6.79691 1.95831 6.72935 1.8552C6.34672 1.27127 5.57156 0.961744 4.86275 1.29386C4.47573 1.4752 4.10709 1.68899 3.76034 1.93166C3.11992 2.37986 3.00031 3.20494 3.31462 3.82814C3.37006 3.93806 3.36467 4.11652 3.23453 4.28317C3.21513 4.30802 3.19591 4.33303 3.17688 4.35819C3.04902 4.52728 2.87761 4.57875 2.75677 4.55359C2.07344 4.41128 1.30647 4.74013 1.0396 5.47589C0.896085 5.87156 0.785336 6.28291 0.710626 6.70658C0.574805 7.4768 1.07423 8.14481 1.7372 8.36325C1.85449 8.40189 1.97652 8.53179 2.00201 8.74071C2.00586 8.77227 2.00996 8.80382 2.01431 8.83533C2.04332 9.04561 1.95837 9.20328 1.85525 9.27085C1.27131 9.6535 0.961775 10.4287 1.29392 11.1375C1.47528 11.5245 1.68908 11.8932 1.93177 12.2399C2.37995 12.8803 3.20496 12.9999 3.82811 12.6856C3.938 12.6302 4.11641 12.6356 4.28302 12.7657C4.308 12.7852 4.33314 12.8045 4.35844 12.8236C4.52747 12.9514 4.57891 13.1228 4.55376 13.2435C4.41147 13.9268 4.7403 14.6937 5.476 14.9605C5.87177 15.1041 6.28322 15.2148 6.707 15.2895C7.4771 15.4253 8.14499 14.926 8.3634 14.2631C8.40202 14.1459 8.53184 14.0239 8.74066 13.9984C8.77245 13.9945 8.8042 13.9904 8.83593 13.986C9.04607 13.957 9.20363 14.0419 9.27113 14.1449C9.65372 14.7288 10.4288 15.0383 11.1375 14.7062C11.5246 14.5248 11.8934 14.3109 12.2402 14.0681C12.8805 13.6199 13.0001 12.7951 12.6858 12.172C12.6304 12.0622 12.6358 11.8838 12.7659 11.7173C12.7854 11.6922 12.8048 11.667 12.8241 11.6416C12.9518 11.4726 13.123 11.4212 13.2437 11.4463C13.9269 11.5886 14.6937 11.2598 14.9605 10.5242C15.1041 10.1284 15.2148 9.71681 15.2896 9.29293C15.4253 8.52287 14.926 7.85503 14.2631 7.63664C14.146 7.59804 14.024 7.46823 13.9985 7.25939C13.9946 7.22765 13.9905 7.19593 13.9861 7.16424C13.9571 6.95411 14.042 6.79657 14.145 6.72907C14.7288 6.34649 15.0383 5.57146 14.7062 4.86277C14.5248 4.47562 14.311 4.10688 14.0682 3.76003C13.62 3.11972 12.7951 3.00014 12.172 3.31441C12.0621 3.36981 11.8837 3.36444 11.7172 3.23436C11.6922 3.21487 11.6671 3.19558 11.6418 3.17647C11.4728 3.04869 11.4213 2.87736 11.4465 2.75661C11.5888 2.07337 11.26 1.30651 10.5243 1.03966C10.1286 0.89611 9.71715 0.785339 9.29339 0.710621ZM8.82401 2.12834C8.87493 1.97377 8.99885 1.92797 9.07634 1.94163C9.42822 2.00368 9.76968 2.09562 10.0981 2.21474C10.1715 2.24136 10.2559 2.34245 10.2228 2.50176C10.0817 3.1792 10.4066 3.80969 10.888 4.1736C10.9081 4.18876 10.928 4.20408 10.9478 4.21956C11.4237 4.59115 12.1169 4.74213 12.7349 4.43049C12.8796 4.35748 12.9991 4.41254 13.0441 4.4768C13.246 4.76523 13.4237 5.07165 13.5743 5.39313C13.6073 5.4636 13.5958 5.59451 13.4599 5.68354C12.8813 6.06271 12.6654 6.73809 12.7479 7.33525C12.7513 7.36043 12.7546 7.38564 12.7577 7.41088C12.831 8.011 13.2151 8.60744 13.872 8.82386C14.0264 8.87475 14.0722 8.99856 14.0585 9.07596C13.9965 9.42794 13.9045 9.7695 13.7854 10.098C13.7588 10.1713 13.6578 10.2557 13.4986 10.2226C12.8212 10.0815 12.1908 10.4065 11.8269 10.8878C11.8117 10.9079 11.7963 10.928 11.7807 10.9479C11.4091 11.4237 11.2581 12.117 11.5697 12.7349C11.6427 12.8796 11.5877 12.9991 11.5234 13.044C11.235 13.2459 10.9286 13.4236 10.6071 13.5743C10.5366 13.6073 10.4057 13.5957 10.3167 13.4598C9.9375 12.8812 9.26215 12.6653 8.665 12.7478C8.63978 12.7513 8.61454 12.7545 8.58927 12.7576C7.98906 12.8309 7.39261 13.215 7.17618 13.8719C7.12528 14.0264 7.00143 14.0722 6.924 14.0585C6.5721 13.9965 6.23062 13.9046 5.9022 13.7854C5.82879 13.7588 5.74432 13.6577 5.7775 13.4984C5.91858 12.821 5.59364 12.1905 5.11232 11.8266C5.09223 11.8114 5.07227 11.796 5.05243 11.7805C4.57656 11.4089 3.88319 11.2578 3.26521 11.5695C3.12043 11.6425 3.00086 11.5875 2.95587 11.5232C2.75405 11.2348 2.57641 10.9285 2.42582 10.6071C2.39274 10.5365 2.40437 10.4055 2.54036 10.3164C3.11906 9.93717 3.33498 9.26173 3.25258 8.6645C3.24913 8.63946 3.24587 8.61438 3.24281 8.58928C3.16955 7.98901 2.78533 7.39248 2.12837 7.17603C1.97378 7.12509 1.92797 7.00116 1.94163 6.92366C2.00367 6.57185 2.0956 6.23045 2.21469 5.90211C2.24134 5.82864 2.34251 5.74414 2.50192 5.77733C3.17942 5.91843 3.80997 5.59345 4.17394 5.11209C4.18905 5.09212 4.2043 5.07226 4.21971 5.05253C4.59136 4.57663 4.74239 3.88324 4.4307 3.26524C4.35766 3.12041 4.41274 3.00079 4.47706 2.95578C4.76541 2.75398 5.07174 2.57635 5.39312 2.42577C5.46368 2.3927 5.59472 2.40433 5.68382 2.54031C6.06304 3.11901 6.73849 3.33492 7.33573 3.25251C7.36075 3.24906 7.38581 3.2458 7.41089 3.24274C8.01112 3.16946 8.60757 2.78525 8.82401 2.12834ZM5.72699 7.99998C5.72699 6.73322 6.7311 5.72705 7.94439 5.72705C8.3561 5.72705 8.74037 5.84149 9.07046 6.04146C9.36568 6.22032 9.75 6.12599 9.92886 5.83076C10.1077 5.53554 10.0134 5.15122 9.71816 4.97236C9.19924 4.65798 8.59193 4.47705 7.94439 4.47705C6.01805 4.47705 4.47699 6.06578 4.47699 7.99998C4.47699 9.93419 6.01805 11.5229 7.94439 11.5229C9.21716 11.5229 10.3242 10.8266 10.9257 9.79966C11.1002 9.50182 11.0002 9.11894 10.7024 8.94447C10.4045 8.77 10.0216 8.87001 9.84717 9.16785C9.45648 9.83481 8.74748 10.2729 7.94439 10.2729C6.7311 10.2729 5.72699 9.26674 5.72699 7.99998ZM11.4284 7.27829C11.298 6.95872 10.9331 6.80541 10.6136 6.93587C10.3018 7.06316 10.1483 7.41354 10.2621 7.72732L10.2621 7.73332C10.2621 8.0785 10.542 8.35832 10.8871 8.35832C11.2323 8.35832 11.5121 8.0785 11.5121 7.73332L11.5122 7.72295C11.5123 7.67271 11.5124 7.58581 11.497 7.49896C11.4797 7.4014 11.4454 7.31911 11.4317 7.28615L11.4284 7.27829Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
